move classes in proper directories
