+++ classes/book/book.rb
