+++ classes/label/label.rb
